Remove asserts and properly reinitialize PyGame display, if needed
The asserts where there to 1. avoid calling run_pygame more than once,
and 2. initialize PyGame display only after SDL_WINDOWID environment
variable was set.

This commit removes the asserts.  1. is solved using a _initialized
flag.  And 2. is solved by checking if the display was set, and using
pygame.display.quit to uninitialize it.