Fix visual lag during startup
Sugar toolbar is rendered before the Pygame canvas.

The deferred initialisation of Pygame is the main cause, and since we
can't easily avoid that, we'll defer drawing of the toolbar instead.

Iconify the activity until the Pygame display is initialised, then
reveal it.

However, this causes breakthrough of the Sugar Home View during startup.