Fix canvas not resizing properly
The configure-event signal is emitted before the canvas size and position are updated.
The size allocation of the PygameCanvas when it occurs is different from the newly
achieved allocation after resizing.
Thus, the parameters sent to it's handler consist of old canvas size, stopping
the canvas from expanding up or shrinking down to it's true allocation.

* Connect _resize_cb handler to size-allocate signal instead of configure-event.